use int

+++ src/main/java/pdc/project/Utils.java
@@ -9,7 +9,7 @@ public final class Utils {
         return new ImageIcon(Objects.requireNonNull(Utils.class.getResource(path))).getImage();
     }
 
-    public static void drawImage(Graphics2D g2d, Image image, double x, double y) {
+    public static void drawImage(Graphics2D g2d, Image image, int x, int y) {
         g2d.drawImage(image, (int) x - image.getWidth(null) / 2, (int) y - image.getHeight(null) / 2, null);
     }
 }
